+++ drizzle/0017_yielding_kylun.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `song` ADD `start_at` real;--> statement-breakpoint
+ALTER TABLE `song` ADD `end_at` real;